Fix InstrumentDefMsg constructor calls for v2-only fields
Fixed issues with InstrumentDefMsg struct construction after adding v2-only fields:

1. V2 decoder: Convert UInt32 to UInt64 when reading raw_instrument_id
   - V2 reads u32 from file but struct expects u64
   - Added UInt64() conversion: `UInt64(read(decoder.io, UInt32))`

2. Test files: Add v2-only fields to constructor calls
   - test_phase2.jl: Added trading_reference_price, trading_reference_date,
     md_security_trading_status, settl_price_type fields
   - test_phase5.jl: Same fields added to serialization test
   - All v2-only fields set to 0 for v3 tests

These changes ensure test InstrumentDefMsg constructor calls match the updated
struct definition with 74 fields instead of the previous 70.

+++ src/decode.jl
@@ -530,7 +530,7 @@ function read_record(decoder::DBNDecoder)
 
             inst_attrib_value = read(decoder.io, Int32)
             underlying_id = read(decoder.io, UInt32)
-            raw_instrument_id = read(decoder.io, UInt32)  # u32 in v2, u64 in v3!
+            raw_instrument_id = UInt64(read(decoder.io, UInt32))  # u32 in v2, convert to u64 for struct
             market_depth_implied = read(decoder.io, Int32)
             market_depth = read(decoder.io, Int32)
             market_segment_id = read(decoder.io, UInt32)
